h3: fix the ControlNet SwiGLU gate/value swap and restore the INT8 MLP route
Two independent defects kept H3 ControlNet from ever completing a denoise.

1. h3_control.cpp reordered `ff.net.0.proj` with a Slice/Slice/Concat and then
   read the result with GateFirst=false, exchanging the SwiGLU gate and value
   halves in all five side blocks. The released Union checkpoint is in the
   DIFFUSERS order, [value; gate], which is already what GateFirst=false wants;
   the base checkpoint's [gate; value] `mlp.fc1` is the opposite. Measured on the
   real checkpoints, control_blocks.0.ff.net.0.proj matches blocks.0.mlp.fc1 only
   ACROSS halves (cos 0.999497/0.999494 crossed, -0.006374/-0.006347 uncrossed).
   Pass the weight through unreordered. QKV needs no change: the split
   attn.to_{q,k,v} concatenated on dim 0 already is the contiguous [Q;K;V] ABI.

   Nothing observable caught this, so h3_control_tests.cpp now pins the operand
   order per side block; the assertion was verified to FAIL when the slice/concat
   is reintroduced and to pass once it is removed.

2. The deployed difc was built from an h3.cpp that had since been reverted, while
   tools/difweights.cpp kept the matching new ABI patterns. Programs it emitted
   carried no GateFirst attribute, so find_h3_w8a8_mlp_plans rejected every base
   chain and admitted 0 of 50. ControlNet is the only path that passes
   --h3-convrot-int8-layers, so it was the only path whose exact-count check
   failed. Reverting difweights.cpp to HEAD and rebuilding difc+difweights
   restores writer/reader agreement: admission goes 0/50 -> 50/50, the 5
   remaining GateFirst rejections are exactly the control side blocks, and the
   sealed denoiser bundle rebinds again (535 bindings, 13 shards).

Also: difh3infer reports the actual and expected chain counts instead of a bare
"not admitted exactly", and cuda_executor gains an env-gated
DIF_MLP_PLAN_DEBUG=1 rejection-reason counter, which is what located (2).

Verified on the 3090 Ti through the dif-inference UI: 1344x768x124 ControlNet
now completes all 6 denoise evaluations and decodes to a 4,049,833-byte MP4 with
synchronized audio.

+++ perf/regress/fixtures/kernel-sources/rotary_frequency.cu
@@ -44,16 +44,18 @@ extern "C" __device__ float dif_silu(float x) {
 }
 // Generic multi-axis rotary tables: each output pair maps to (axis,
 // component); omega = 1 / (theta * ntk) ^ (2 * component / axis_dim).
+// Keep F64 intermediates, as in the CPU semantic implementation and creator
+// rope(). Only the completed tables round to F32; early rounding changes Q/K.
 extern "C" __global__ void dif_op_1(const dif_f32* positions, const int* pair_axes, const int* pair_indices, const int* axis_dims, dif_f32* cosine, dif_f32* sine) {
   unsigned long long i = (unsigned long long)blockIdx.x * blockDim.x + threadIdx.x;
   if (i < 12ULL) {
     unsigned long long pair = i % 3ULL, token = (i / 3ULL) % 4ULL,
                        batch = i / (3ULL * 4ULL);
     int axis = pair_axes[pair], component = pair_indices[pair], axis_dim = axis_dims[axis];
-    float scale = (2.0f * (float)component) / (float)axis_dim;
-    float omega = 1.0f / powf((float)(10000 * 1), scale);
-    float angle = dif_load_f32(positions, (batch * 4ULL + token) * 2ULL + (unsigned long long)axis) * omega;
-    dif_store_f32(cosine, i, cosf(angle));
-    dif_store_f32(sine, i, sinf(angle));
+    double scale = (2.0 * (double)component) / (double)axis_dim;
+    double omega = 1.0 / pow((double)10000 * (double)1, scale);
+    double angle = (double)dif_load_f32(positions, (batch * 4ULL + token) * 2ULL + (unsigned long long)axis) * omega;
+    dif_store_f32(cosine, i, (float)cos(angle));
+    dif_store_f32(sine, i, (float)sin(angle));
   }
 }
